pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.32.0",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "pygments>=2.19.2",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
    "pyttsx3>=2.99",
    "speechrecognition>=3.14.3",
]
